feat: add reference decoders to soothe
Add decoders such as the h264, h265 and
av1 reference decoders to decode content
after encoding.
This will help to validate that the encoder
is the only changing brick for the pipeline.

The decoders does not support Y4M as output format
so use a YUV to Y4M converter to generate the
files.

It will use reference decoder from the fraunhofer
institute for h26x and Open Media for AV1.